add wavelength- or time- dependent `ok` masks
